lab: remove restrictive details managed by host provider
The host provider of the lab is in charge of mandatory compliance
and technical details to make machines work as expected.
The provider has a better knowledge of what should be done
so we can remove those requirements.
This way the requirements are more oriented to the results
seen as customers than the how.

Signed-off-by: Thomas Monjalon <thomas@monjalon.net>